MeetWrap1/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.2
huggingface-hub==0.19.4
